changed ports, and reworked delete when there is an exception
still using port 3000, but not hard coding it. also used the .findByIdAndUpdate() method to see if it is a valid id

+++ backend/controllers/product.controller.js
@@ -15,7 +15,7 @@ export const getProducts =  async (req, res) => {
 export const createProduct = async (req, res) => {
     const product = req.body; // user will send this data body
     if (!product.name || !product.price || !product.image) {
-        return res.status(400).json({ success:false, message: "Please provide all fields" });
+        return res.status(404).json({ success:false, message: "Please provide all fields" });
     }
 
     const newProduct = new Product(product)
@@ -50,11 +50,16 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async(req, res) => {
     const {id} = req.params;
+
+    if (!product.name || !product.price || !product.image) {
+        return res.status(404).json({ success:false, message: "Please provide all fields" });
+    }
+
     try{
         await Product.findByIdAndDelete(id);
         res.status(200).json({success: true, message: "Product deleted"});
     }
     catch(error) {
-        res.status(404).json({success: false, message: "Product not found"});
+        res.status(500).json({success: false, message: "Server Error"});
     }
 }
